requirement.txt
streamlit
numpy
pandas
yfinance
requests
google-generativeai
beautifulsoup4
fpdf
